Refactor/server defaults (#40)
* feat: add index.ts to defaults folder

* refactor: rename import @defaults/[...] to @defaults

* chore(tsconfig): repname @defaults/* path to @defaults
