+++ server/src/controllers/configsController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { getConfigs, updateConfigs, ConfigUpdateData } from "@services";
-import { configDefaults } from "@defaults/configsDefaults";
+import { configDefaults } from "@defaults";
 
 export const getConfigsList = async (req: Request, res: Response, next: NextFunction) => {
   try {
